horizontal player movement
Implemented a way to handle horizontal player movement with "A" and "D"; experimented with capping the fps

+++ src/GameObject.cpp
@@ -2,7 +2,7 @@
 #include "TextureManager.h"
 
 // how we create our object
-GameObject::GameObject(int x, int y, int width, int height, const char* texturesheet, SDL_Renderer* ren)
+GameObject::GameObject(float x, float y, int width, int height, const char* texturesheet, SDL_Renderer* ren)
   : renderer_(ren), x_(x), y_(y), width_(width), height_(height) {
   objTexture_ = TextureManager::LoadTexture(texturesheet, ren);
 }
@@ -17,7 +17,7 @@ void GameObject::Update(float delta_time) {
 
 void GameObject::Render() {
   if (objTexture_ && renderer_) {
-    SDL_Rect dstRect = {x_, y_, width_, height_}; // default size
+    SDL_Rect dstRect = {static_cast<int>(x_), static_cast<int>(y_), width_, height_}; // default size
 
     SDL_RenderCopy(renderer_, objTexture_, nullptr, &dstRect);
   }
