+++ src/GameObject.h
@@ -4,22 +4,22 @@
 
 class GameObject {
   public:
-    GameObject(int x, int y, int width, int height, const char* texturesheet, SDL_Renderer* ren);
+    GameObject(float x, float y, int width, int height, const char* texturesheet, SDL_Renderer* ren);
     virtual ~GameObject();
 
     virtual void Update(float delta_time);
     virtual void Render();
 
-    int GetX() const { return x_; }
-    int GetY() const { return y_; }
+    float GetX() const { return x_; }
+    float GetY() const { return y_; }
     int GetHeight() const { return height_; }
     int GetWidth() const { return width_; }
 
-    void SetX(int newX) { x_ = newX; }
-    void SetY(int newY) { y_ = newY; }
+    void SetX(float newX) { x_ = newX; }
+    void SetY(float newY) { y_ = newY; }
     
   private: 
-    int x_, y_;
+    float x_, y_;
     int height_, width_;
     SDL_Texture* objTexture_;
     SDL_Renderer* renderer_;
